+++ resvg-qt/src/lib.rs
@@ -1,7 +1,6 @@
 use std::ffi::CString;
 use std::i32;
 use std::ops::{Deref, DerefMut};
-use std::path::PathBuf;
 use std::slice;
 
 #[allow(dead_code)]
@@ -84,16 +83,6 @@ impl Image {
 
     pub fn new_rgba_premultiplied(width: u32, height: u32) -> Option<Image> {
         unsafe { Self::from_ptr(ffi::qtc_qimage_create_rgba_premultiplied(width, height)) }
-    }
-
-    pub fn from_file(path: &PathBuf) -> Option<Image> {
-        let c_path = CString::new(path.to_str().unwrap()).unwrap();
-
-        unsafe { Self::from_ptr(ffi::qtc_qimage_from_file(c_path.as_ptr())) }
-    }
-
-    pub fn from_data(data: &[u8]) -> Option<Image> {
-        unsafe { Self::from_ptr(ffi::qtc_qimage_from_data(data.as_ptr(), data.len() as i32)) }
     }
 
     unsafe fn from_ptr(img: *mut ffi::qtc_qimage) -> Option<Image> {