+++ src/backend_cairo/mod.rs
@@ -4,11 +4,12 @@
 
 //! Cairo backend implementation.
 
-use cairo::MatrixTrait;
+use usvg::try_opt;
 use log::warn;
 
-use crate::{prelude::*, layers};
-use crate::backend_utils::ConvTransform;
+use crate::prelude::*;
+use crate::layers::{self, Layer, Layers};
+use crate::backend_utils::{self, FlatRender, ConvTransform, BlendMode};
 
 
 macro_rules! try_create_surface {
@@ -26,14 +27,10 @@ macro_rules! try_create_surface {
 }
 
 
-mod clip_and_mask;
 mod filter;
 mod image;
 mod path;
 mod style;
-
-
-type CairoLayers = layers::Layers<cairo::ImageSurface>;
 
 
 impl ConvTransform<cairo::Matrix> for usvg::Transform {
@@ -183,17 +180,15 @@ pub fn render_node_to_canvas(
     img_size: ScreenSize,
     cr: &cairo::Context,
 ) {
-    let mut layers = create_layers(img_size);
+    let tree = node.tree();
+    let mut render = CairoFlatRender::new(&tree, opt, img_size, cr);
 
-    apply_viewbox_transform(view_box, img_size, &cr);
-
-    let curr_ts = cr.get_matrix();
     let mut ts = node.abs_transform();
     ts.append(&node.transform());
 
-    cr.transform(ts.to_native());
-    render_node(node, opt, &mut layers, cr);
-    cr.set_matrix(curr_ts);
+    render.apply_viewbox(view_box, img_size);
+    render.apply_transform(ts);
+    render.render_node(node);
 }
 
 fn create_surface(
@@ -207,160 +202,196 @@ fn create_surface(
     Some((surface, img_size))
 }
 
-/// Applies viewbox transformation to the painter.
-fn apply_viewbox_transform(
-    view_box: usvg::ViewBox,
-    img_size: ScreenSize,
-    cr: &cairo::Context,
-) {
-    let ts = utils::view_box_to_transform(view_box.rect, view_box.aspect, img_size.to_size());
-    cr.transform(ts.to_native());
-}
-
-fn render_node(
-    node: &usvg::Node,
-    opt: &Options,
-    layers: &mut CairoLayers,
-    cr: &cairo::Context,
-) -> Option<Rect> {
-    match *node.borrow() {
-        usvg::NodeKind::Svg(_) => {
-            render_group(node, opt, layers, cr)
+impl Into<cairo::Operator> for BlendMode {
+    fn into(self) -> cairo::Operator {
+        match self {
+            BlendMode::SourceOver => cairo::Operator::Over,
+            BlendMode::Clear => cairo::Operator::Clear,
+            BlendMode::DestinationIn => cairo::Operator::DestIn,
+            BlendMode::DestinationOut => cairo::Operator::DestOut,
+            BlendMode::Xor => cairo::Operator::Xor,
         }
-        usvg::NodeKind::Path(ref path) => {
-            path::draw(&node.tree(), path, opt, cr)
-        }
-        usvg::NodeKind::Image(ref img) => {
-            Some(image::draw(img, opt, cr))
-        }
-        usvg::NodeKind::Group(ref g) => {
-            render_group_impl(node, g, opt, layers, cr)
-        }
-        _ => None,
     }
 }
 
-fn render_group(
-    parent: &usvg::Node,
-    opt: &Options,
-    layers: &mut CairoLayers,
-    cr: &cairo::Context,
-) -> Option<Rect> {
-    let curr_ts = cr.get_matrix();
-    let mut g_bbox = Rect::new_bbox();
-
-    for node in parent.children() {
-        cr.transform(node.transform().to_native());
-
-        let bbox = render_node(&node, opt, layers, cr);
-        if let Some(bbox) = bbox {
-            if let Some(bbox) = bbox.transform(&node.transform()) {
-                g_bbox = g_bbox.expand(bbox);
-            }
-        }
-
-        // Revert transform.
-        cr.set_matrix(curr_ts);
+impl layers::Image for cairo::ImageSurface {
+    fn new(size: ScreenSize) -> Option<Self> {
+        Some(try_create_surface!(size, None))
     }
 
-    // Check that bbox was changed, otherwise we will have a rect with x/y set to f64::MAX.
-    if g_bbox.fuzzy_ne(&Rect::new_bbox()) {
-        Some(g_bbox)
-    } else {
-        None
-    }
-}
-
-fn render_group_impl(
-    node: &usvg::Node,
-    g: &usvg::Group,
-    opt: &Options,
-    layers: &mut CairoLayers,
-    cr: &cairo::Context,
-) -> Option<Rect> {
-    let sub_surface = layers.get()?;
-    let mut sub_surface = sub_surface.borrow_mut();
-
-    let curr_ts = cr.get_matrix();
-
-    let bbox = {
-        let sub_cr = cairo::Context::new(&*sub_surface);
-        sub_cr.set_matrix(curr_ts);
-
-        render_group(node, opt, layers, &sub_cr)
-    };
-
-    // Filter can be rendered on an object without a bbox,
-    // as long as filter uses `userSpaceOnUse`.
-    if let Some(ref id) = g.filter {
-        if let Some(filter_node) = node.tree().defs_by_id(id) {
-            if let usvg::NodeKind::Filter(ref filter) = *filter_node.borrow() {
-                let ts = usvg::Transform::from_native(&curr_ts);
-                filter::apply(filter, bbox, &ts, opt, &mut *sub_surface);
-            }
-        }
-    }
-
-    // Clipping and masking can be done only for objects with a valid bbox.
-    if let Some(bbox) = bbox {
-        if let Some(ref id) = g.clip_path {
-            if let Some(clip_node) = node.tree().defs_by_id(id) {
-                if let usvg::NodeKind::ClipPath(ref cp) = *clip_node.borrow() {
-                    let sub_cr = cairo::Context::new(&*sub_surface);
-                    sub_cr.set_matrix(curr_ts);
-
-                    clip_and_mask::clip(&clip_node, cp, opt, bbox, layers, &sub_cr);
-                }
-            }
-        }
-
-        if let Some(ref id) = g.mask {
-            if let Some(mask_node) = node.tree().defs_by_id(id) {
-                if let usvg::NodeKind::Mask(ref mask) = *mask_node.borrow() {
-                    let sub_cr = cairo::Context::new(&*sub_surface);
-                    sub_cr.set_matrix(curr_ts);
-
-                    clip_and_mask::mask(&mask_node, mask, opt, bbox, layers, &sub_cr);
-                }
-            }
-        }
-    }
-
-    let curr_matrix = cr.get_matrix();
-    cr.set_matrix(cairo::Matrix::identity());
-    cr.set_source_surface(&*sub_surface, 0.0, 0.0);
-    if !g.opacity.is_default() {
-        cr.paint_with_alpha(g.opacity.value());
-    } else {
+    fn clear(&mut self) {
+        let cr = cairo::Context::new(self);
+        cr.set_operator(cairo::Operator::Clear);
+        cr.set_source_rgba(0.0, 0.0, 0.0, 0.0);
         cr.paint();
     }
-
-    cr.set_matrix(curr_matrix);
-
-    // All layers must be unlinked from the main context/cr after used.
-    // TODO: find a way to automate this
-    cr.reset_source_rgba();
-
-    bbox
 }
 
-fn create_layers(
-    img_size: ScreenSize,
-) -> CairoLayers {
-    layers::Layers::new(img_size, create_subsurface, clear_subsurface)
+struct CairoFlatRender<'a> {
+    tree: &'a usvg::Tree,
+    opt: &'a Options,
+    cr: &'a cairo::Context,
+    layers: Layers<cairo::ImageSurface>,
 }
 
-fn create_subsurface(
-    size: ScreenSize,
-) -> Option<cairo::ImageSurface> {
-    Some(try_create_surface!(size, None))
+impl<'a> CairoFlatRender<'a> {
+    fn new(
+        tree: &'a usvg::Tree,
+        opt: &'a Options,
+        img_size: ScreenSize,
+        cr: &'a cairo::Context,
+    ) -> Self {
+        CairoFlatRender {
+            tree,
+            opt,
+            cr,
+            layers: Layers::new(img_size),
+        }
+    }
+
+    fn new_painter(layer: &mut Layer<cairo::ImageSurface>) -> cairo::Context {
+        let cr = cairo::Context::new(&mut layer.img);
+        cr.set_matrix(layer.ts.to_native());
+        cr.set_operator(layer.blend_mode.into());
+
+        if let Some(rect) = layer.clip_rect {
+            cr.rectangle(rect.x(), rect.y(), rect.width(), rect.height());
+            cr.clip();
+        }
+
+        cr
+    }
+
+    fn paint<F>(&mut self, f: F)
+        where F: FnOnce(&usvg::Tree, &Options, &cairo::Context)
+    {
+        match self.layers.current_mut() {
+            Some(layer) => {
+                let cr = Self::new_painter(layer);
+                f(self.tree, self.opt, &cr);
+            }
+            None => {
+                f(self.tree, self.opt, self.cr);
+            }
+        }
+    }
 }
 
-fn clear_subsurface(
-    surface: &mut cairo::ImageSurface,
-) {
-    let cr = cairo::Context::new(&surface);
-    cr.set_operator(cairo::Operator::Clear);
-    cr.set_source_rgba(0.0, 0.0, 0.0, 0.0);
-    cr.paint();
+impl<'a> FlatRender for CairoFlatRender<'a> {
+    fn draw_path(&mut self, path: &usvg::Path, bbox: Option<Rect>) {
+        self.paint(|tree, opt, cr| {
+            path::draw(tree, path, opt, bbox, cr);
+        });
+    }
+
+    fn draw_svg_image(&mut self, image: &usvg::Image) {
+        self.paint(|_, opt, cr| {
+            image::draw_svg(&image.data, image.view_box, opt, cr);
+        });
+    }
+
+    fn draw_raster_image(&mut self, image: &usvg::Image) {
+        self.paint(|_, opt, cr| {
+            image::draw_raster(
+                image.format, &image.data, image.view_box, image.rendering_mode, opt, cr,
+            );
+        });
+    }
+
+    fn filter(&mut self, filter: &usvg::Filter, bbox: Option<Rect>, ts: usvg::Transform) {
+        if let Some(layer) = self.layers.current_mut() {
+            filter::apply(filter, bbox, &ts, &self.opt, &mut layer.img);
+        }
+    }
+
+    fn fill_layer(&mut self, r: u8, g: u8, b: u8, a: u8) {
+        if let Some(layer) = self.layers.current_mut() {
+            let clip_cr = cairo::Context::new(&layer.img);
+            clip_cr.set_source_rgba(
+                r as f64 / 255.0,
+                g as f64 / 255.0,
+                b as f64 / 255.0,
+                a as f64 / 255.0,
+            );
+            clip_cr.paint();
+        }
+    }
+
+    fn push_layer(&mut self) -> Option<()> {
+        self.layers.push()
+    }
+
+    fn pop_layer(&mut self, opacity: usvg::Opacity, mode: BlendMode) {
+        let last = try_opt!(self.layers.pop());
+        match self.layers.current_mut() {
+            Some(prev) => {
+                let cr = cairo::Context::new(&prev.img);
+                cr.set_source_surface(&last.img, 0.0, 0.0);
+                cr.set_operator(mode.into());
+                if !opacity.is_default() {
+                    cr.paint_with_alpha(opacity.value());
+                } else {
+                    cr.paint();
+                }
+            }
+            None => {
+                let curr_matrix = self.cr.get_matrix();
+                self.reset_transform();
+                self.cr.set_source_surface(&last.img, 0.0, 0.0);
+                self.cr.set_operator(mode.into());
+                if !opacity.is_default() {
+                    self.cr.paint_with_alpha(opacity.value());
+                } else {
+                    self.cr.paint();
+                }
+
+                self.cr.set_matrix(curr_matrix);
+                self.cr.set_operator(cairo::Operator::Over);
+
+                // All layers must be unlinked from the main context/cr after used.
+                self.cr.reset_source_rgba();
+            }
+        }
+
+        self.layers.push_back(last);
+    }
+
+    fn apply_mask(&mut self) {
+        let img_size = self.layers.img_size();
+        if let Some(layer) = self.layers.current_mut() {
+            let mut data = try_opt!(layer.img.get_data().ok());
+            backend_utils::image_to_mask(&mut data, img_size);
+        }
+    }
+
+    fn set_composition_mode(&mut self, mode: BlendMode) {
+        match self.layers.current_mut() {
+            Some(layer) => layer.blend_mode = mode,
+            None => self.cr.set_operator(mode.into()),
+        }
+    }
+
+    fn set_clip_rect(&mut self, rect: Rect) {
+        match self.layers.current_mut() {
+            Some(layer) => layer.clip_rect = Some(rect),
+            None => {
+                self.cr.rectangle(rect.x(), rect.y(), rect.width(), rect.height());
+                self.cr.clip();
+            }
+        }
+    }
+
+    fn get_transform(&self) -> usvg::Transform {
+        match self.layers.current() {
+            Some(layer) => layer.ts,
+            None => usvg::Transform::from_native(&self.cr.get_matrix()),
+        }
+    }
+
+    fn set_transform(&mut self, ts: usvg::Transform) {
+        match self.layers.current_mut() {
+            Some(layer) => layer.ts = ts,
+            None => self.cr.set_matrix(ts.to_native()),
+        }
+    }
 }