+++ src/backend_qt/style.rs
@@ -5,7 +5,9 @@
 use crate::qt;
 use usvg::try_opt;
 
-use crate::{prelude::*, backend_utils::ConvTransform};
+use crate::prelude::*;
+use crate::backend_utils::{ConvTransform, FlatRender};
+use super::QtFlatRender;
 
 
 pub fn fill(
@@ -225,9 +227,10 @@ fn prepare_pattern(
         p.scale(bbox.width(), bbox.height());
     }
 
-    let mut layers = super::create_layers(img_size);
-    super::render_group(pattern_node, opt, &mut layers, &mut p);
-    p.end();
+    let ref tree = pattern_node.tree();
+    let mut render = QtFlatRender::new(tree, opt, img_size, &mut p);
+    render.render_group(pattern_node);
+    render.finish();
 
     let img = if !opacity.is_default() {
         // If `opacity` isn't `1` then we have to make image semitransparent.