+++ tools/usvg/src/main.rs
@@ -8,7 +8,7 @@ use std::io::{self, Read, Write};
 use std::path::Path;
 use std::process;
 
-use gumdrop::Options;
+use pico_args::Arguments;
 
 use usvg::svgdom;
 
@@ -65,88 +65,84 @@ ARGS:
 ");
 }
 
-#[derive(Debug, Options)]
+#[derive(Debug)]
 struct Args {
-    #[options()]
     help: bool,
-
-    #[options(short = "V")]
     version: bool,
-
-    #[options(short = "c", no_long)]
     stdout: bool,
-
-    #[options(no_short)]
     keep_named_groups: bool,
-
-    #[options(no_short, meta = "DPI", default = "96", parse(try_from_str = "parse_dpi"))]
     dpi: u32,
-
-    #[options(no_short, meta = "FAMILY", default = "Times New Roman")]
     font_family: String,
-
-    #[options(no_short, meta = "SIZE", default = "12", parse(try_from_str = "parse_font_size"))]
     font_size: u32,
-
-    #[options(no_short, meta = "LANG", parse(try_from_str = "parse_languages"))]
-    languages: Option<Vec<String>>,
-
-    #[options(no_short, meta = "HINT", default = "geometricPrecision", parse(try_from_str))]
+    languages: Vec<String>,
     shape_rendering: usvg::ShapeRendering,
-
-    #[options(no_short, meta = "HINT", default = "optimizeLegibility", parse(try_from_str))]
     text_rendering: usvg::TextRendering,
-
-    #[options(no_short, meta = "HINT", default = "optimizeQuality", parse(try_from_str))]
     image_rendering: usvg::ImageRendering,
-
-    #[options(no_short, meta = "INDENT", default = "4", parse(try_from_str = "parse_indent"))]
     indent: svgdom::Indent,
-
-    #[options(no_short, meta = "INDENT", default = "none", parse(try_from_str = "parse_indent"))]
     attrs_indent: svgdom::Indent,
-
-    #[options(no_short)]
     quiet: bool,
-
-    #[options(free)]
     free: Vec<String>,
 }
 
-fn parse_dpi(s: &str) -> Result<u32, &'static str> {
+fn collect_args() -> Result<Args, pico_args::Error> {
+    let mut input = Arguments::from_env();
+    Ok(Args {
+        help:               input.contains("--help"),
+        version:            input.contains(["-V", "--version"]),
+        stdout:             input.contains("-c"),
+        keep_named_groups:  input.contains("--keep-named-groups"),
+        dpi:                input.value_from_fn("--dpi", parse_dpi)?.unwrap_or(96),
+        font_family:        input.value_from_str("--font-family")?
+                                 .unwrap_or_else(|| "Times New Roman".to_string()),
+        font_size:          input.value_from_fn("--font-size", parse_font_size)?.unwrap_or(12),
+        languages:          input.value_from_fn("--languages", parse_languages)?
+                                 .unwrap_or(vec!["en".to_string()]), // TODO: use system language
+        shape_rendering:    input.value_from_str("--shape-rendering")?.unwrap_or_default(),
+        text_rendering:     input.value_from_str("--text-rendering")?.unwrap_or_default(),
+        image_rendering:    input.value_from_str("--image-rendering")?.unwrap_or_default(),
+        indent:             input.value_from_fn("--indent", parse_indent)?
+                                 .unwrap_or(svgdom::Indent::Spaces(4)),
+        attrs_indent:       input.value_from_fn("--attrs-indent", parse_indent)?
+                                 .unwrap_or(svgdom::Indent::None),
+        quiet:              input.contains("--quiet"),
+        free:               input.free()?,
+    })
+}
+
+fn parse_dpi(s: &str) -> Result<u32, String> {
     let n: u32 = s.parse().map_err(|_| "invalid number")?;
 
     if n >= 10 && n <= 4000 {
         Ok(n)
     } else {
-        Err("DPI out of bounds")
+        Err("DPI out of bounds".to_string())
     }
 }
 
-fn parse_font_size(s: &str) -> Result<u32, &'static str> {
+fn parse_font_size(s: &str) -> Result<u32, String> {
     let n: u32 = s.parse().map_err(|_| "invalid number")?;
 
     if n > 0 && n <= 192 {
         Ok(n)
     } else {
-        Err("font size out of bounds")
+        Err("font size out of bounds".to_string())
     }
 }
 
-fn parse_languages(s: &str) -> Result<Vec<String>, &'static str> {
+fn parse_languages(s: &str) -> Result<Vec<String>, String> {
     let mut langs = Vec::new();
     for lang in s.split(',') {
         langs.push(lang.trim().to_string());
     }
 
     if langs.is_empty() {
-        return Err("languages list cannot be empty");
+        return Err("languages list cannot be empty".to_string());
     }
 
     Ok(langs)
 }
 
-fn parse_indent(s: &str) -> Result<svgdom::Indent, &'static str> {
+fn parse_indent(s: &str) -> Result<svgdom::Indent, String> {
     let indent = match s {
         "none" => svgdom::Indent::None,
         "0" => svgdom::Indent::Spaces(0),
@@ -155,7 +151,7 @@ fn parse_indent(s: &str) -> Result<svgdom::Indent, &'static str> {
         "3" => svgdom::Indent::Spaces(3),
         "4" => svgdom::Indent::Spaces(4),
         "tabs" => svgdom::Indent::Tabs,
-        _ => return Err("invalid INDENT value"),
+        _ => return Err("invalid INDENT value".to_string()),
     };
 
     Ok(indent)
@@ -175,8 +171,7 @@ enum OutputTo<'a> {
 
 
 fn main() {
-    let args: Vec<String> = ::std::env::args().collect();
-    let args = match Args::parse_args_default(&args[1..]) {
+    let args = match collect_args() {
         Ok(v) => v,
         Err(e) => {
             eprintln!("Error: {}.", e);
@@ -238,11 +233,6 @@ fn process(args: &Args) -> Result<(), String> {
         (svg_from, svg_to)
     };
 
-    let languages = match args.languages.as_ref() {
-        Some(v) => v.clone(),
-        None => vec!["en".to_string()], // TODO: use system language
-    };
-
     let re_opt = usvg::Options {
         path: match in_svg {
             InputFrom::Stdin => None,
@@ -251,7 +241,7 @@ fn process(args: &Args) -> Result<(), String> {
         dpi: args.dpi as f64,
         font_family: args.font_family.clone(),
         font_size: args.font_size as f64,
-        languages,
+        languages: args.languages.clone(),
         shape_rendering: args.shape_rendering,
         text_rendering: args.text_rendering,
         image_rendering: args.image_rendering,